feat(presentation/search): move search to your files and add Filter model to modal
