+++ package.json
@@ -28,6 +28,7 @@
     "@react-navigation/stack": "^6.3.20",
     "axios": "^0.27.2",
     "mobx": "^6.12.0",
+    "mobx-react": "^9.1.0",
     "react": "18.0.0",
     "react-dom": "^18.2.0",
     "react-native": "0.69.2",
