+++ src/main/factories/presentation/searchFactory.tsx
@@ -1,8 +1,23 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import React from 'react';
 import { RouteProp } from '@react-navigation/native';
 import { Routes } from '~/main/navigation';
-import Search from '../../../../src/presentation/search';
+import { Filter, Ordination } from '~/domain/enums';
+import { SearchPlaces } from '~/domain/useCases';
+import { SearchPlaceModel } from '~/domain/models';
+import { FilterParam } from '~/domain/params';
 import { StackParams } from '../../../../src/main/navigation/navigation';
+import { Search, useSearch } from '../../../../src/presentation/search';
+
+class SearchPlacesFake implements SearchPlaces {
+  async search(
+    place: string,
+    { type, ordination }: FilterParam,
+    nextPageToken?: string | undefined,
+  ): Promise<SearchPlaceModel[]> {
+    return [];
+  }
+}
 
 type Props = {
   route: RouteProp<StackParams, Routes>;
@@ -10,16 +25,12 @@ type Props = {
 };
 
 const SearchFactory = ({}: Props) => {
-  return (
-    <Search
-      searchValue={''}
-      changeSearch={() => {}}
-      searchTo={() => {}}
-      filter={() => {}}
-      showFilterOptions={false}
-      placeList={[]}
-    />
-  );
+  const viewModel = useSearch({
+    filterParam: { ordination: Ordination.Closer, type: Filter.Entertainment },
+    nextPageToken: '',
+    searchPlaces: new SearchPlacesFake(),
+  });
+  return <Search {...viewModel} />;
 };
 
 export default SearchFactory;
