+++ src/main/stores/index.ts
@@ -1,0 +1,13 @@
+import Filter from '../../presentation/search/model/Filter';
+
+class RootStore {
+  filter: Filter;
+
+  constructor() {
+    this.filter = new Filter();
+  }
+}
+
+const store = new RootStore();
+
+export default store;
